requirements.txt
PyQt5 
pillow